Switch to global communication buffers
With these arrays, it seems you cannot pass in slices to the TRGTOL
subroutines. Those slices are then passed to MPI_SEND/RECV, and that
seemed to cause an issue. It's better to pass the full arrays, along
with the appropriate offset indices. This also ensures the compiler
doesn't create a local copy to store the slice.